+++ src/seyond_lidar_ros/src/multi_fusion/ros1_multi_fusion.hpp
@@ -10,7 +10,7 @@
 #include <message_filters/subscriber.h>
 #include <message_filters/sync_policies/approximate_time.h>
 #include <message_filters/synchronizer.h>
-#include <pcl_ros/point_cloud.h>
+#include <pcl_conversions/pcl_conversions.h>
 
 
 #include <vector>