+++ src/seyond_lidar_ros/src/test/ros1_test.hpp
@@ -10,7 +10,6 @@
 #pragma once
 #include <ros/ros.h>
 #include <sensor_msgs/PointCloud2.h>
-#include <pcl_conversions/pcl_conversions.h>
 
 #include <limits>
 #include <memory>