+++ src/seyond_lidar_ros/src/driver/ros1_driver_adapter.hpp
@@ -11,6 +11,7 @@
 #include <pcl_conversions/pcl_conversions.h>
 #include <ros/ros.h>
 #include <sensor_msgs/PointCloud2.h>
+#include <sensor_msgs/Imu.h>
 #include <std_msgs/Float64.h>
 
 #include <limits>
@@ -51,6 +52,12 @@ class ROSAdapter {
       driver_ptr_->register_publish_packet_callback(std::bind(&ROSAdapter::publishPacket, this, std::placeholders::_1,
                                                               std::placeholders::_2, std::placeholders::_3,
                                                               std::placeholders::_4));
+    }
+
+    if (lidar_config_.enable_imu_msg) {
+      inno_imu_pub_ = nh_->advertise<sensor_msgs::Imu>("/iv_imu", 10);
+      driver_ptr_->register_publish_imu_callback(
+          std::bind(&ROSAdapter::publishImu, this, std::placeholders::_1, std::placeholders::_2));
     }
   }
 
@@ -126,6 +133,19 @@ class ROSAdapter {
     inno_frame_pub_.publish(std::move(ros_msg));
   }
 
+  void publishImu(std::vector<float> &imu_data, double timestamp) {
+    sensor_msgs::Imu ros_msg;
+    ros_msg.header.frame_id = lidar_config_.frame_id;
+    ros_msg.header.stamp = ros::Time().fromSec(timestamp * 1e-6);
+    ros_msg.linear_acceleration.x = imu_data[0];
+    ros_msg.linear_acceleration.y = imu_data[1];
+    ros_msg.linear_acceleration.z = imu_data[2];
+    ros_msg.angular_velocity.x = imu_data[3];
+    ros_msg.angular_velocity.y = imu_data[4];
+    ros_msg.angular_velocity.z = imu_data[5];
+    inno_imu_pub_.publish(std::move(ros_msg));
+  }
+
  private:
   seyond::LidarConfig lidar_config_;
   std::shared_ptr<ros::NodeHandle> nh_;
@@ -134,6 +154,7 @@ class ROSAdapter {
 
   ros::Publisher inno_frame_pub_;
   ros::Publisher inno_pkt_pub_;
+  ros::Publisher inno_imu_pub_;
   ros::Subscriber inno_pkt_sub_;
 
   std::unique_ptr<seyond::SeyondScan> inno_scan_msg_;
@@ -206,6 +227,7 @@ class ROSNode {
     private_nh_->param("reflectance_mode", lidar_config.reflectance_mode, true);
     private_nh_->param("multiple_return", lidar_config.multiple_return, 1);
     private_nh_->param("enable_falcon_ring", lidar_config.enable_falcon_ring, false);
+    private_nh_->param("enable_imu_msg", lidar_config.enable_imu_msg, false);
 
     private_nh_->param("continue_live", lidar_config.continue_live, false);